pyproject.toml
[tool.poetry]
name = "drf_auth"
version = "0.1.0"
description = ""
authors = ["Haneen Hashlamoun <haneen.hashlamoun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
httpie = "^2.6.0"
django-cors-headers = "^3.10.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
